requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
alembic
python-Levenshtein
geoalchemy2
pytest
